fix: compile errors — navController param, formatCount Long, LastFmProvider import, JSONObject import
Five compile errors fixed:

1. LastFMSettings.kt:333 — 'Unresolved reference navController'
   LastFmSettingsSuccess() didn't have navController as a parameter.
   Added it to both the function signature and the call site.

2. LastFmDashboardScreen.kt:527 — 'Argument type mismatch: Number & Comparable vs Long'
   info.playcount is Int? but formatCount() expects Long. Added .toLong().

3. LastFmLoginScreen.kt:28 — 'Unresolved reference LastFMProvider'
   The enum class is named LastFmProvider (mixed case), not LastFMProvider.
   Fixed the import.

4. SourceCheckService.kt:184 — 'Unresolved reference JSONObject'
   Missing import org.json.JSONObject. Added it.

5. SourceCheckService.kt:204 — 'Unresolved reference head'
   Request.Builder.head() not available in this OkHttp version. Replaced
   with .method("HEAD", null).

+++ app/src/main/kotlin/moe/rukamori/archivetune/ui/screens/settings/LastFmDashboardScreen.kt
@@ -524,7 +524,7 @@ private fun UserCard(
                             ) {
                                 Column(horizontalAlignment = Alignment.CenterHorizontally) {
                                     Text(
-                                        text = formatCount(info.playcount ?: 0L),
+                                        text = formatCount((info.playcount ?: 0).toLong()),
                                         style = MaterialTheme.typography.displaySmall,
                                         fontWeight = FontWeight.Bold,
                                         color = MaterialTheme.colorScheme.onPrimaryContainer,
